ota-ubi.bbclass: add a check whether update_ubi.zip is created
fail safely with an warning message if ota zip file generation failed

+++ classes/ota-ubi.bbclass
@@ -169,6 +169,10 @@ do_gen_otazip_ubi[dirs] += "${DEPLOY_DIR_IMAGE}/ota-scripts"
 do_gen_otazip_ubi() {
     ./full_ota.sh ${OTA_TARGET_FILES_UBI_PATH} ${IMAGE_ROOTFS} ubi --system_path ${IMAGE_SYSTEM_MOUNT_POINT}
 
-    cp update_ubi.zip ${DEPLOY_DIR_IMAGE}/${IMAGE_BASENAME}
+    if [[ -e update_ubi.zip ]]; then
+        cp update_ubi.zip ${DEPLOY_DIR_IMAGE}/${IMAGE_BASENAME}
+    else
+        bbwarn "update_ubi.zip failed to create"
+    fi
 }
 addtask do_gen_otazip_ubi after do_recovery_ubi before do_build
